Move bar chart utility methods into view holder
